Move state testing into separate entry file
